feat(setting): 页面配置-主题色
